requirements.txt
fastapi==0.66.0
numpy==1.19.5
scikit-learn==0.24.1
tensorflow==2.5.0rc2
pydantic==1.8.2
pyspellchecker==0.6.2
nltk==3.6.2
git+https://github.com/ClaudeCoulombe/FrenchLefffLemmatizer.git
